Fix MultiBinding Equality comparison and GetHashCode to hash bound resources. Move test MultiConverters to the Demo project.
