+++ Forge.Forms/src/Forge.Forms.Demo/Infrastructure/DemoAppController.cs
@@ -2,7 +2,9 @@ using Forge.Application.Infrastructure;
 using Forge.Application.Routing;
 using Forge.Forms.Controls;
 using Forge.Forms.Demo.Behaviors;
+using Forge.Forms.Demo.Converters;
 using Forge.Forms.Demo.Routes;
+using Forge.Forms.DynamicExpressions;
 
 namespace Forge.Forms.Demo.Infrastructure
 {
@@ -10,6 +12,10 @@ namespace Forge.Forms.Demo.Infrastructure
     {
         protected override void OnInitializing()
         {
+            // Add some multiconverters for testing
+            Resource.MultiValueConverters["Divide"] = new DivideMultiConverter();
+            Resource.MultiValueConverters["Multiply"] = new MultiplyMultiConverter();
+
             DynamicForm.AddBehavior(new CheckAllBehavior());
             var factory = Routes.RouteFactory;
