+++ Forge.Forms/src/Forge.Forms/DynamicExpressions/MultiBinding.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using System.Windows.Data;
 using Forge.Forms.FormBuilding;
 
@@ -41,9 +42,20 @@ namespace Forge.Forms.DynamicExpressions
         {
             if (other is MultiBinding resource)
             {
-                return Resources == resource.Resources
-                       && OneTimeBinding == resource.OneTimeBinding
-                       && ValueConverter == resource.ValueConverter;
+                if (Resources.Length == resource.Resources.Length
+                    && OneTimeBinding == resource.OneTimeBinding
+                    && ValueConverter == resource.ValueConverter)
+                {
+                    for (int i = 0; i < Resources.Length; i++)
+                    {
+                        if (Resources[i].Equals(resource.Resources[i]) == false)
+                        {
+                            return false;
+                        }
+                    }
+
+                    return true;
+                }
             }
 
             return false;
@@ -51,7 +63,13 @@ namespace Forge.Forms.DynamicExpressions
 
         public override int GetHashCode()
         {
-            return Resources.GetHashCode() ^ (OneTimeBinding ? 123456789 : 741852963);
+            var hashCode = (OneTimeBinding ? 123456789 : 741852963);
+            foreach (var resource in Resources)
+            {
+                hashCode = resource.GetHashCode() ^ hashCode;
+            }
+
+            return hashCode;
         }
     }
 }
