+++ Forge.Forms/src/Forge.Forms/Forge.Forms.csproj
@@ -117,8 +117,6 @@
     <Compile Include="DynamicExpressions\MultiValueBinding.cs" />
     <Compile Include="DynamicExpressions\InvariantDouble.cs" />
     <Compile Include="DynamicExpressions\InvariantInt.cs" />
-    <Compile Include="DynamicExpressions\MultiValueConverters\DivideMultiConverter.cs" />
-    <Compile Include="DynamicExpressions\MultiValueConverters\MultiplyMultiConverter.cs" />
     <Compile Include="DynamicExpressions\ValueConverters\IsEqualConverter.cs" />
     <Compile Include="FormBuilding\Defaults\DirectContentField.cs" />
     <Compile Include="FormBuilding\Defaults\Properties\DirectContentBuilder.cs" />
@@ -424,5 +422,6 @@
       <Generator>XamlIntelliSenseFileGenerator</Generator>
     </Page>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
